requirements.txt
Django
djangorestframework>=3.14.0
django-taggit>=3.0.0
Pillow>=9.5.0
#psycopg2>=2.9.6  Use this for PostgreSQL (or remove if using SQLite)
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-filter>=23.2
gunicorn>=21.2.0  # Required for deploying on Heroku or similar platforms
# whitenoise>=6.5.0  # Static file handling for production
python-decouple>=3.8
sentry-sdk>=2.25.1